backend/requirements.txt
fastapi[all]==0.110.0
sqlalchemy[asyncio]==2.0.28
alembic==1.13.1
asyncpg==0.29.0
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.9
cryptography==42.0.5
uvicorn[standard]==0.29.0
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
